requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
streamlit>=1.37.0
pandas>=2.2.2
openai>=1.35.0
autogen-agentchat>=0.3.0
httpx>=0.27.0

